refactor: extract dependencies-versions-resolver code to a new file
