+++ src/consumer/component/dependencies/dependency-resolver/index.js
@@ -1,0 +1,4 @@
+// @flow
+
+export { default as loadDependenciesForComponent } from './dependencies-resolver';
+export { default as updateDependenciesVersions } from './dependencies-versions-resolver';
