+++ src/consumer/consumer.js
@@ -36,10 +36,7 @@ import { MissingBitMapComponent } from './bit-map/exceptions';
 import logger from '../logger/logger';
 import DirStructure from './dir-structure/dir-structure';
 import { getLatestVersionNumber, pathNormalizeToLinux } from '../utils';
-import {
-  loadDependenciesForComponent,
-  updateDependenciesVersions
-} from './component/dependencies/dependency-resolver/dependencies-resolver';
+import { loadDependenciesForComponent, updateDependenciesVersions } from './component/dependencies/dependency-resolver';
 import { Version, Component as ModelComponent } from '../scope/models';
 import MissingFilesFromComponent from './component/exceptions/missing-files-from-component';
 import ComponentNotFoundInPath from './component/exceptions/component-not-found-in-path';
